+++ CHANGELOG.md
@@ -17,6 +17,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## Changed
 
 - Upgraded Python build/test system
+- Switched to `argparse`
 
 # 0.3.2 - 2021-01-30